requirements.txt
# Core Web Framework
Flask>=2.0
Werkzeug==2.0.3
Jinja2==3.1.2

# WebSocket Support
Flask-SocketIO==5.3.6
python-socketio==5.3.0
python-engineio==4.1.0
gevent==25.5.1
gevent-websocket==0.10.1 # Required for Gunicorn SocketIO worker
websocket-client>=1.6.0  # Required for TV runtime Socket.IO client (WebSocket transport)
gunicorn==22.0.0         # Production WSGI server for Linux/macOS
dnspython==2.3.0

# Media Processing
numpy<2.0.0

# Utilities
pyperclip==1.8.2  # Clipboard operations for sharing URLs
requests==2.28.2  # HTTP requests for testing and API operations
psutil==5.9.5    # System monitoring and process management
pyudev==0.24.1   # Device monitoring (HDMI hotplug)
MarkupSafe==2.1.5
qrcode[pil]==7.4.2 # For Wireguard QR code generation
PyYAML==6.0.1      # For Headscale config
specter-runtime    # SPECTER backend runtime

# GhostStream Transcoding Client (optional - for transcoding support)
ghoststream==1.2.5 # Official GhostStream Python SDK
httpx>=0.24.0      # Also used directly by GhostHub API/proxy code
websockets>=12.0   # WebSocket client for GhostHub real-time progress updates

# Packaging and Distribution
pyinstaller==6.13.0  # For building standalone executable

# Testing
pytest>=7.0.0        # Test framework
pytest-cov>=4.0.0    # Code coverage reporting
pytest-mock>=3.10.0  # Mock fixtures for pytest
